model_training/requirements.txt
# Model Training - Dependencies
# Install with: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
google-cloud-storage>=2.10.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
